+++ yandex/data_source_yandex_alb_target_group.go
@@ -49,11 +49,15 @@ func dataSourceYandexALBTargetGroup() *schema.Resource {
 					Schema: map[string]*schema.Schema{
 						"subnet_id": {
 							Type:     schema.TypeString,
-							Required: true,
+							Optional: true,
 						},
 						"ip_address": {
 							Type:     schema.TypeString,
 							Required: true,
+						},
+						"private_ipv4_address": {
+							Type:     schema.TypeBool,
+							Optional: true,
 						},
 					},
 				},